LayoutTests/fast/xmlhttprequest/xmlhttprequest-sync-disabled-expected.txt

CONSOLE MESSAGE: Synchronous XMLHttpRequests cannot be made in the current window context.
This tests that synchronous XMLHttpRequests fail when they are disabled for documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.open("GET", "http://mydomain/", false) threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS successfullyParsed is true

TEST COMPLETE
